etc: 회원가입, 로그인, 초기세팅 페이지 directory 변경
