+++ dressmeup/lib/components/SIgnin/Sign_up.dart
@@ -1,15 +1,18 @@
 import 'package:dressmeup/components/init_home.dart';
-import 'package:dressmeup/components/init_setting.dart';
+import 'package:dressmeup/components/SIgnin/init_setting.dart';
 import 'package:flutter/material.dart';
 import 'package:dressmeup/components/changeModel/change_model.dart';
 
 /// Flutter code sample for [Radio].
 
-class Login extends StatelessWidget {
-  Login({super.key});
+class SignUp extends StatelessWidget {
+  SignUp({super.key});
 
   final idController = TextEditingController();
   final pwController = TextEditingController();
+  final cpwController = TextEditingController();
+  final nickController = TextEditingController();
+  final phoneController = TextEditingController();
 
   @override
   Widget build(BuildContext context) {
@@ -39,9 +42,36 @@ class Login extends StatelessWidget {
               Padding(
                 padding: const EdgeInsets.all(8.0),
                 child: TextField(
+                  obscureText: true,
                   controller: pwController,
                   decoration: const InputDecoration(
                       border: OutlineInputBorder(), labelText: 'Password'),
+                ),
+              ),
+              Padding(
+                padding: const EdgeInsets.all(8.0),
+                child: TextField(
+                  obscureText: true,
+                  controller: cpwController,
+                  decoration: const InputDecoration(
+                      border: OutlineInputBorder(),
+                      labelText: 'Check Password'),
+                ),
+              ),
+              Padding(
+                padding: const EdgeInsets.all(8.0),
+                child: TextField(
+                  controller: nickController,
+                  decoration: const InputDecoration(
+                      border: OutlineInputBorder(), labelText: 'Nickname'),
+                ),
+              ),
+              Padding(
+                padding: const EdgeInsets.all(8.0),
+                child: TextField(
+                  controller: phoneController,
+                  decoration: const InputDecoration(
+                      border: OutlineInputBorder(), labelText: 'Phone Number'),
                 ),
               ),
               const SizedBox(height: 40),
@@ -57,7 +87,7 @@ class Login extends StatelessWidget {
                   },
                   child: Padding(
                     padding: const EdgeInsets.all(8.0),
-                    child: Text('Login',
+                    child: Text('SignUp',
                         style: TextStyle(
                             fontSize: 25, color: Colors.grey.shade700)),
                   )),
