+++ dressmeup/lib/components/SIgnin/login.dart
@@ -1,18 +1,15 @@
 import 'package:dressmeup/components/init_home.dart';
-import 'package:dressmeup/components/init_setting.dart';
+import 'package:dressmeup/components/SIgnin/init_setting.dart';
 import 'package:flutter/material.dart';
 import 'package:dressmeup/components/changeModel/change_model.dart';
 
 /// Flutter code sample for [Radio].
 
-class SignUp extends StatelessWidget {
-  SignUp({super.key});
+class Login extends StatelessWidget {
+  Login({super.key});
 
   final idController = TextEditingController();
   final pwController = TextEditingController();
-  final cpwController = TextEditingController();
-  final nickController = TextEditingController();
-  final phoneController = TextEditingController();
 
   @override
   Widget build(BuildContext context) {
@@ -42,36 +39,9 @@ class SignUp extends StatelessWidget {
               Padding(
                 padding: const EdgeInsets.all(8.0),
                 child: TextField(
-                  obscureText: true,
                   controller: pwController,
                   decoration: const InputDecoration(
                       border: OutlineInputBorder(), labelText: 'Password'),
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.all(8.0),
-                child: TextField(
-                  obscureText: true,
-                  controller: cpwController,
-                  decoration: const InputDecoration(
-                      border: OutlineInputBorder(),
-                      labelText: 'Check Password'),
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.all(8.0),
-                child: TextField(
-                  controller: nickController,
-                  decoration: const InputDecoration(
-                      border: OutlineInputBorder(), labelText: 'Nickname'),
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.all(8.0),
-                child: TextField(
-                  controller: phoneController,
-                  decoration: const InputDecoration(
-                      border: OutlineInputBorder(), labelText: 'Phone Number'),
                 ),
               ),
               const SizedBox(height: 40),
@@ -87,7 +57,7 @@ class SignUp extends StatelessWidget {
                   },
                   child: Padding(
                     padding: const EdgeInsets.all(8.0),
-                    child: Text('SignUp',
+                    child: Text('Login',
                         style: TextStyle(
                             fontSize: 25, color: Colors.grey.shade700)),
                   )),
